+++ app/helpers/amp_html/base_helper.rb
@@ -10,15 +10,20 @@ module AmpHtml
             end
         end
 
-        # Fix for accessing AMP assets in different environments
-        def get_asset name
-            if Rails.application.assets
-                asset = Rails.application.assets[name]
-                return asset.to_s if asset
-            end
-            asset = Rails.application.assets_manifest.assets[name]
-            return nil unless asset
-            return File.binread(File.join(Rails.application.assets_manifest.dir, asset))
+        def amp_packaged_css
+            global = Rails.application.assets.find_asset('amp/application.css').to_s
+            package = Rails.application.assets.find_asset("amp/packages/#{params[:controller]}/#{action_name}.css").to_s
+            # if Rails.env.development?
+            #     ::Sass::Engine.new(global + package, {
+            #         syntax: :scss,
+            #         cache: false,
+            #         read_cache: false,
+            #         style: :compressed
+            #     }).render
+            # else
+            #     global + package
+            # end
+            global + package
         end
 
         def amp= value = true