+++ lib/amp-html/configuration.rb
@@ -11,12 +11,10 @@ module AmpHtml
 
     class Configuration
 
-        attr_accessor :inherit_stylesheets
         attr_accessor :split_view
         attr_accessor :split_view_default
 
         def initialize
-            @inherit_stylesheets = false
             @split_view = false
             @split_view_default = 'amp'
         end