+++ lib/generators/templates/initializer.rb
@@ -1,9 +1,5 @@
 AmpHtml.configure do |config|
 
-    # Use your regular stylesheets bundled with `app/assets/stylesheets/application.css` in your AMP views. Takes a boolean.
-	# config.inherit_stylesheets = false
-
-
     ### SplitView
 
 	# Enable SplitView to allow for AMP enabled and AMP disabled versions of your views. Takes a boolean.